Title

Title

intCC(
  data,
  method = NULL,
  individualK = NULL,
  individualMaxK = 10,
  globalK = NULL,
  B = 1000,
  pItem = 0.8,
  pFeature = 1,
  ccClMethods = "kmeans",
  ccDistHCs = "euclidean",
  hclustMethod = "average",
  finalclmethod = "hclust",
  finalhclustMethod = "average",
  Silhouette = TRUE,
  widestGap = FALSE,
  dunns = FALSE,
  dunn2s = FALSE
)

Arguments

data

List of M datasets, each of size N X P_m, m = 1, ..., M.

method

1 layer or 2 layer.

individualK

Vector containing the number of clusters in each dataset. Default is NULL. If the number of clusters is not provided, then all the possible values between 2 and individualMaxK are considered and the best value is chosen for each dataset by maximising the silhouette.

individualMaxK

Maximum number of clusters considered for the individual data. Default is 10.

globalK

Number of global clusters. Default is NULL. If the number of clusters is not provided, then all the number of clusters for each dataset will be used and decided according to criterion defined.

B

Number of iterations for consensus clustering. Default is 1000.

pItem

Proportion of items sampled at each iteration.

pFeature

Proportion of features sampled at each iteration.

ccClMethods

The i-th element of this vector goes into the "clMethod" argument of consensusCluster() for the i-th dataset. If only one string is provided, then the same method is used for all datasets.

ccDistHCs

Distance used for each clustering method. Default is Euclidean distance.

hclustMethod

Method used for hclust method only worked when hclust method used in consensus clustering.

finalclmethod

Method used for final clustering. It could be pam or hclust.

finalhclustMethod

Method used for hclust method only worked when hclust method used in final clustering.

Silhouette

Boolean. If TRUE, compute Silhouette to choose best number of clusters. Default is TRUE.

widestGap

Boolean. If TRUE, compute also widest gap index to choose best number of clusters. Default is FALSE.

dunns

Boolean. If TRUE, compute also Dunn's index to choose best number of clusters. Default is FALSE.

dunn2s

Boolean. If TRUE, compute also alternative Dunn's index to choose best number of clusters. Default is FALSE.

Value

clustering results, weights and weighted consensus matrix

Examples

library(intCC)
load(system.file("extdata", "exampleData.RData", package = "intCC"))
res <- intCC(exampleData, method="1 layer", individualK = rep(3,4),
globalK = 3, pFeature = 0.8 ,ccClMethods = "kmeans",
ccDistHCs = "euclidean",hclustMethod = "average",finalclmethod="hclust",
finalhclustMethod = "average",Silhouette=TRUE)
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations
#> Warning: did not converge in 1000 iterations